requirements.txt
aiohttp
brawlstats
clashroyale
discord-webhook
git+https://github.com/Rapptz/discord.py@rewrite#egg=discord.py[voice]
getgist
lxml
flask